updates variant calling isdocker; module load

+++ variants/somatic_vc.sh
@@ -34,8 +34,11 @@ do
     esac
 done
 
-source /etc/profile.d/modules.sh
-module load htslib/gcc/1.8 samtools/gcc/1.8 snpeff/4.3q vcftools/0.1.14
+if [[ -z $isdocker ]]
+then
+    source /etc/profile.d/modules.sh
+    module load htslib/gcc/1.8 samtools/gcc/1.8 snpeff/4.3q vcftools/0.1.14
+fi
 export PATH=/project/shared/bicf_workflow_ref/seqprg/bin:$PATH
 
 shift $(($OPTIND -1))
@@ -83,7 +86,10 @@ fi
 
 if [ $algo == 'strelka2' ]
 then
-    module load strelka/2.9.10 manta/1.3.1 
+    if [[ -z $isdocker ]]
+    then
+	module load strelka/2.9.10 manta/1.3.1
+    fi
     opt=''
     if [[ -n $tbed ]]
     then
@@ -110,7 +116,10 @@ then
     vcf-concat strelka/results/variants/*.vcf.gz | vcf-annotate -n --fill-type -n |vcf-sort |java -jar $SNPEFF_HOME/SnpSift.jar filter "(GEN[*].DP >= 10)" | perl -pe "s/TUMOR/${tid}/g" | perl -pe "s/NORMAL/${nid}/g" |bgzip > ${pair_id}.strelka2.vcf.gz
 elif [ $algo == 'virmid' ]
 then 
-    module load virmid/1.2
+    if [[ -z $isdocker ]]
+    then
+	module load virmid/1.2
+    fi
     cosmic=${index_path}/cosmic.vcf.gz
     if [[ ! -f "${index_path}/cosmic.vcf.gz" ]]
     then
@@ -120,20 +129,29 @@ then
     virmid -R ${reffa} -D ${tumor} -N ${normal} -s ${cosmic} -t $NPROC -M 2000 -c1 10 -c2 10
     perl $baseDir/addgt_virmid.pl ${tumor}.virmid.som.passed.vcf
     perl $baseDir/addgt_virmid.pl ${tumor}.virmid.loh.passed.vcf
-    module rm java/oracle/jdk1.7.0_51
-    module load snpeff/4.3q
+    if [[ -z $isdocker ]]
+    then
+	module rm java/oracle/jdk1.7.0_51
+	module load snpeff/4.3q
+    fi
     vcf-concat *gt.vcf | vcf-sort | vcf-annotate -n --fill-type -n | java -jar $SNPEFF_HOME/SnpSift.jar filter '((NDP >= 10) & (DDP >= 10))' | perl -pe "s/TUMOR/${tid}/g" | perl -pe "s/NORMAL/${nid}/g" | bgzip > ${pair_id}.virmid.vcf.gz
 elif [ $algo == 'mutect' ]
 then
-    module load gatk/4.1.4.0 parallel/20150122
+    if [[ -z $isdocker ]]
+    then
+	module load gatk/4.1.4.0 parallel/20150122
+    fi
     threads=`expr $NPROC / 2`
     gatk --java-options "-Xmx20g" Mutect2 $ponopt  --independent-mates -RF AllowAllReadsReadFilter -R ${reffa} -I ${tumor} -tumor ${tid} -I ${normal} -normal ${nid} --output ${tid}.mutect.vcf -L $interval
     vcf-concat ${tid}.mutect.*vcf | vcf-sort | vcf-annotate -n --fill-type | java -jar $SNPEFF_HOME/SnpSift.jar filter -p '(GEN[*].DP >= 10)' | bgzip > ${pair_id}.mutect.vcf.gz
 elif [ $algo == 'varscan' ]
 then
-    module load bcftools/gcc/1.8 VarScan/2.4.2
-    module rm java/oracle/jdk1.7.0_51
-    module load snpeff/4.3q 
+    if [[ -z $isdocker ]]
+    then
+	module load bcftools/gcc/1.8 VarScan/2.4.2
+	module rm java/oracle/jdk1.7.0_51
+	module load snpeff/4.3q
+    fi
     samtools mpileup -C 50 -f ${reffa} $tumor > t.mpileup
     samtools mpileup -C 50 -f ${reffa} $normal > n.mpileup
     VarScan somatic n.mpileup t.mpileup vscan --output-vcf 1
@@ -141,13 +159,14 @@ then
     vcf-concat vscan*.vcf | vcf-sort | vcf-annotate -n --fill-type -n | java -jar $SNPEFF_HOME/SnpSift.jar filter '((exists SOMATIC) & (GEN[*].DP >= 10))' | perl -pe "s/TUMOR/${tid}/" | perl -pe "s/NORMAL/${nid}/g" | bgzip >  ${pair_id}.varscan.vcf.gz
 elif [ $algo == 'shimmer' ]
 then
-    module load R/3.6.1-gccmkl
-    module rm perl/5.18.2
+    if [[ -z $isdocker ]]
+    then
+	module load R/3.6.1-gccmkl
+	module rm perl/5.18.2
+    fi
     shimmer.pl --minqual 25 --ref ${reffa} ${normal} ${tumor} --outdir shimmer 2> shimmer.err
     perl $baseDir/add_readct_shimmer.pl
     module rm java/oracle/jdk1.7.0_51
     module load snpeff/4.3q
     vcf-annotate -n --fill-type shimmer/somatic_diffs.readct.vcf | java -jar $SNPEFF_HOME/SnpSift.jar filter '(GEN[*].DP >= 10)' | perl -pe "s/TUMOR/${tid}/" | perl -pe "s/NORMAL/${nid}/g" | bgzip > ${pair_id}.shimmer.vcf.gz
 fi
-
-
